# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fathom game API
 * OpenAPI spec version: 0.1.0
 */
import type { ExposureStatus } from './exposureStatus';
import type { GameStatePhase } from './gameStatePhase';
import type { GameStateStatus } from './gameStateStatus';
import type { GameStateWinCondition } from './gameStateWinCondition';
import type { OpponentPieceState } from './opponentPieceState';
import type { PieceState } from './pieceState';
import type { ProximityReveal } from './proximityReveal';

export interface GameState {
  id: string;
  status: GameStateStatus;
  /**
     * 1 or 2, null if spectating
     * @nullable
     */
  yourNumber: number | null;
  /** @nullable */
  yourName: string | null;
  /** @nullable */
  opponentName: string | null;
  /**
     * 1 or 2, whose turn to move
     * @nullable
     */
  currentTurnPlayer: number | null;
  phase: GameStatePhase;
  round: number;
  /** Perspective rows (1-6) exposed by monsoon — visible to both players */
  exposedRows: number[];
  exposure: ExposureStatus;
  yourPieces: PieceState[];
  opponentPieces: OpponentPieceState[];
  proximityHistory: ProximityReveal[];
  /** True if the mover has committed their move this turn */
  moveCommitted: boolean;
  isYourTurn: boolean;
  /** True when viewing in fog-free spectator mode (all pieces revealed) */
  spectating?: boolean;
  /** @nullable */
  winner: number | null;
  /** @nullable */
  winCondition: GameStateWinCondition;
}